Update user.model.ts
I encountered an issue creating a new user when trying to save the credentials. It kept telling me that the field 'user_id' does not exit on the model. I found out that the problem hast do do with the fact that the 'keyTo' was referencing the wrong field after changing it to 'userId' it worked just fine.